Time: 2 ms (100.00%) | Memory: 44.6 MB (22.45%) - LeetSync

+++ 437-path-sum-iii/path-sum-iii.java
@@ -17,29 +17,26 @@
 import java.util.HashMap;
 
 class Solution {
-    private int DFS(TreeNode node, long currSum, int target, HashMap<Long, Long> prefixSumCount){
+    private int DFS(TreeNode node, long currSum, int target, HashMap<Long, Integer> prefixSumCount){
         if(node == null) return 0;
         currSum+=node.val;
 
-        // Get the number of possible paths
-        long ctr = prefixSumCount.getOrDefault(currSum-target, 0L);
+        // Get the number of possible paths ending at the current node.
+        int ctr = prefixSumCount.getOrDefault(currSum-target, 0);
 
-        prefixSumCount.put(currSum, prefixSumCount.getOrDefault(currSum, 0L)+1L);
+        prefixSumCount.put(currSum, prefixSumCount.getOrDefault(currSum, 0)+1);
 
         ctr += DFS(node.left, currSum, target, prefixSumCount);
         ctr += DFS(node.right, currSum, target, prefixSumCount);
 
-        prefixSumCount.put(currSum, prefixSumCount.get(currSum)-1L);
-        // if (prefixSumCount.get(currSum) == 0) {
-        //     prefixSumCount.remove(currSum);
-        // }
+        prefixSumCount.put(currSum, prefixSumCount.get(currSum)-1);
 
-        return (int)ctr;
+        return ctr;
     }
 
     public int pathSum(TreeNode root, int targetSum) {
-        HashMap<Long, Long> prefixSumCount = new HashMap<>();
-        prefixSumCount.put(0L, 1L);
+        HashMap<Long, Integer> prefixSumCount = new HashMap<>();
+        prefixSumCount.put(0L, 1);
         return DFS(root, 0L, targetSum, prefixSumCount);
     }
 }
